+++ library/src/scripts/components/PageHeading.tsx
@@ -11,7 +11,7 @@ import Heading from "@library/components/Heading";
 
 interface IPageHeading {
     title: string;
-    backUrl?: string; // back link
+    backUrl?: string | null; // back link
     className?: string;
     menu?: JSX.Element;
 }
