+++ library/src/scripts/components/modal/ModalSizes.ts
@@ -1,0 +1,12 @@
+/**
+ * @author Stéphane LaFlèche <stephane.l@vanillaforums.com>
+ * @copyright 2009-2018 Vanilla Forums Inc.
+ * @license GPL-2.0-only
+ */
+
+export enum ModalSizes {
+    FULL_SCREEN = "full screen",
+    LARGE = "large",
+    MEDIUM = "medium",
+    SMALL = "small",
+}
