+++ library/src/scripts/components/modal/index.ts
@@ -5,6 +5,7 @@
  */
 
 export { default as Modal } from "./Modal";
+export { ModalSizes } from "./ModalSizes";
 export { default as ModalLink } from "./ModalLink";
 export { default as ModalLoader } from "./ModalLoader";
 export { default as ModalRouter } from "./ModalRouter";
